requirements.txt
# Kafka
kafka-python>=2.0.2

# Data validation
pydantic>=2.0

# Date/time utilities
python-dateutil>=2.8

# Testing
pytest>=7.0
pytest-asyncio>=0.21

# Development
python-dotenv>=1.0
